move vocab_jax to the original file
