+++ vllm/model_executor/models/qwen2.py
@@ -433,8 +433,9 @@ class Qwen2Model(nnx.Module):
             ("gate_up_proj", "gate_proj", 0),
             ("gate_up_proj", "up_proj", 1),
         ]
-        params_dict = nnx.state(self, nnx.Param)
-        print("params_dict:", params_dict.keys())
+        params_dict = nnx.state(self)
+        with open("../param_dict.txt", "w") as f:
+            f.write(str(params_dict.keys()))
         loaded_params: set[str] = set()
         for name, loaded_weight in weights:
             print("name:", name)
